requirements.txt
kafka-python
opencv-python
Flask
